Create FileTree
